MS-80 refactor(slider): code refactoring for main page content and slider for main page
